requirements.txt
httpx
streamlit
python-dotenv
google-cloud-texttospeech
azure-cognitiveservices-speech
elevenlabs
pyht
